services/application-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic[email]==2.6.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dateutil==2.9.0
